material_estimate/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pdfplumber==0.7.6
pytesseract==0.3.13
Pillow==12.0.0
openai>=1.0.0
python-dotenv==1.0.0